+++ src/i18n/i18n.config.ts
@@ -14,6 +14,11 @@ import ptbr from './locales/pt-BR.json';
 import tr from './locales/tr.json';
 import bn from './locales/bn.json';
 import id from './locales/id.json';
+import nl from './locales/nl.json';
+import nb from './locales/nb.json';
+import bg from './locales/bg.json';
+import gl from './locales/gl.json';
+import cs from './locales/cs.json';
 
 export default defineI18nConfig(() => ({
   legacy: false,
@@ -35,5 +40,10 @@ export default defineI18nConfig(() => ({
     tr,
     bn,
     id,
+    nl,
+    nb,
+    bg,
+    gl,
+    cs,
   },
 }));