+++ src/server/database/repositories/interface/types.ts
@@ -38,18 +38,19 @@ export const InterfaceUpdateSchema = schemaForType<InterfaceUpdateType>()(
     s2: SSchema,
     s3: SSchema,
     s4: SSchema,
+    h1: HSchema,
+    h2: HSchema,
+    h3: HSchema,
+    h4: HSchema,
     i1: ISchema,
     i2: ISchema,
     i3: ISchema,
     i4: ISchema,
     i5: ISchema,
-    h1: HSchema,
-    h2: HSchema,
-    h3: HSchema,
-    h4: HSchema,
     port: PortSchema,
     device: device,
     enabled: EnabledSchema,
+    firewallEnabled: EnabledSchema,
   })
 );