+++ src/server/routes/metrics/prometheus.get.ts
@@ -39,8 +39,6 @@ async function getPrometheusResponse() {
   const id = `interface="${wgInterface.name}"`;
 
   const returnText = [
-    '# HELP wg-easy and wireguard metrics',
-    '',
     '# HELP wireguard_configured_peers',
     '# TYPE wireguard_configured_peers gauge',
     `wireguard_configured_peers{${id}} ${wireguardPeerCount}`,